\name{ConsPlan-package}
\alias{ConsPlan-package}
\alias{ConsPlan}
\docType{package}
\title{
Conservation Planning Tools
}
\description{
This is a package that will be used for conservation planning exercises.
}
\details{
\tabular{ll}{
Package: \tab ConsPlan\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-04-30\cr
License: \tab GPL version 2 or newer\cr
LazyLoad: \tab yes\cr
}

}
\author{
Jeremy VanDerWal and Stephanie Januchowski.

Maintainer: Jeremy VanDerWal \email{jjvanderwal@gmail.com}
}

\keyword{ package }

